prediccion-7/requirements.txt
Flask>=3.0.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
